Q4/2022 release, total of 393 code examples

+++ code_examples/MTU_MBIST_1_KIT_TC397_TFT/Libraries/iLLD/TC39B/Tricore/Eray/Eray/IfxEray_Eray.h
@@ -3,8 +3,8 @@
  * \brief ERAY ERAY details
  * \ingroup IfxLld_Eray
  *
- * \version iLLD_1_0_1_12_0
- * \copyright Copyright (c) 2019 Infineon Technologies AG. All rights reserved.
+ * \version iLLD_1_0_1_15_0_1
+ * \copyright Copyright (c) 2021 Infineon Technologies AG. All rights reserved.
  *
  *
  *                                 IMPORTANT NOTICE
@@ -676,7 +676,7 @@ typedef struct
     uint8                  collisionAvoidanceDuration;       /**< \brief accepted duration of collision avoidance symbol. */
     IfxEray_StrobePosition strobePosition;                   /**< \brief sample count value for strobing. */
     IfxEray_Baudrate       baudrate;                         /**< \brief baud rate on the flexray bus. */
-    uint8                  receiveWakeupTestDuration;        /**< \brief duration of received wakeup pattern in bit times. */
+    uint16                 receiveWakeupTestDuration;        /**< \brief duration of received wakeup pattern in bit times. */
     uint8                  transmitWakeupRepetitions;        /**< \brief duration of transmitted wakeup pattern in bit times. */
     uint16                 transmissionStartTime;            /**< \brief duration of transmission start time. */
 } IfxEray_Eray_Prtc1Control;
@@ -734,7 +734,7 @@ typedef struct
 {
     uint32                    networkVectorLength;           /**< \brief length of network management vector. */
     uint8                     staticFramepayload;            /**< \brief payload length of static frame in double bytes. */
-    uint8                     latestTransmissionStart;       /**< \brief number of dynamic slots before inhibit frame transmission in dynamic segment. */
+    uint16                    latestTransmissionStart;       /**< \brief number of dynamic slots before inhibit frame transmission in dynamic segment. */
     IfxEray_Eray_GTUConfig    gtuConfig;                     /**< \brief gloabl timing unit configuration structure. */
     IfxEray_Eray_Succ1Config  succ1Config;                   /**< \brief communication controller control properties. */
     IfxEray_Eray_Succ2Config  succ2Config;                   /**< \brief communication listen timeout properties. */
